Move refactored app tests next to file its testing
We try to colocate tests where possible.
